[FE Feat]불필요한 코드 삭제 및 피드백 반영
